requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
config==0.5.1
croniter==5.0.1
discord==2.3.2
discord-paginator==0.3.1
discord-py-interactions==5.13.2
discord-py-slash-command==4.2.1
discord-typings==0.9.0
discord.py==2.4.0
discordlists.py==1.1.1
DiscordUtils==1.3.4
emoji==2.14.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
parsedatetime==2.6
pendulum==3.0.0
pillow==11.0.0
propcache==0.2.1
psutil==6.1.0
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
time-machine==2.16.0
timeago==1.0.16
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
yarl==1.18.3
